[android][WEEX-187] weex custom font not rendered right when font is first downloaded
